[INJICERT-13] removed vci service impl module and refactored code Signed-off-by: Challarao <challarao@beehyv.com>
