+++ certify-core/pom.xml
@@ -20,35 +20,15 @@
     <description>Certify Core Library</description>
 
     <properties>
-      <springdoc-openapi-ui-version>1.6.9</springdoc-openapi-ui-version>
-        <lombok.version>1.18.30</lombok.version>
+        <springdoc-openapi-webmvc-ui-version>2.5.0</springdoc-openapi-webmvc-ui-version>
         <commons.validator.version>1.7</commons.validator.version>
     </properties>
 
     <dependencies>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-validation</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.springdoc</groupId>
-            <artifactId>springdoc-openapi-ui</artifactId>
-            <version>${springdoc-openapi-ui-version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <scope>provided</scope>
-            <version>${lombok.version}</version>
+            <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+            <version>${springdoc-openapi-webmvc-ui-version}</version>
         </dependency>
         <dependency>
             <groupId>commons-validator</groupId>
