+++ certify-core/src/main/java/io/mosip/certify/core/dto/CredentialDefinition.java
@@ -1,4 +1,4 @@
-package io.mosip.certify.core.dto.vci;
+package io.mosip.certify.core.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
