+++ certify-core/src/main/java/io/mosip/certify/core/dto/ParsedAccessToken.java
@@ -3,10 +3,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package io.mosip.certify.core.dto.vci;
+package io.mosip.certify.core.dto;
 
 import lombok.Data;
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
