--- certify-core/src/main/java/io/mosip/certify/core/dto/vci/VCIssuanceTransaction.java
@@ -1,16 +1,0 @@
-package io.mosip.certify.core.dto.vci;
-
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-@Data
-public class VCIssuanceTransaction implements Serializable {
-
-    private String cNonce;
-    private long cNonceIssuedEpoch;
-    private int cNonceExpireSeconds;
-
-
-}
