+++ certify-service/pom.xml
@@ -28,18 +28,8 @@
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>io.mosip.certify</groupId>
             <artifactId>certify-core</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.mosip.certify</groupId>
-            <artifactId>vci-service-impl</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
