+++ certify-service/src/main/java/io/mosip/certify/filter/AccessTokenValidationFilter.java
@@ -3,11 +3,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package io.mosip.certify.vci.filter;
+package io.mosip.certify.filter;
 
 import io.mosip.certify.core.constants.Constants;
-import io.mosip.certify.core.dto.vci.ParsedAccessToken;
-import io.mosip.certify.core.util.IdentityProviderUtil;
+import io.mosip.certify.core.dto.ParsedAccessToken;
+import io.mosip.certify.core.util.CommonUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -31,16 +31,16 @@ import java.util.Objects;
 @Component
 public class AccessTokenValidationFilter extends OncePerRequestFilter {
 
-    @Value("${mosip.certify.vci.authn.issuer-uri}")
+    @Value("${mosip.certify.authn.issuer-uri}")
     private String issuerUri;
 
-    @Value("${mosip.certify.vci.authn.jwk-set-uri}")
+    @Value("${mosip.certify.authn.jwk-set-uri}")
     private String jwkSetUri;
 
-    @Value("#{${mosip.certify.vci.authn.allowed-audiences}}")
+    @Value("#{${mosip.certify.authn.allowed-audiences}}")
     private List<String> allowedAudiences;
 
-    @Value("#{${mosip.certify.vci.authn.filter-urls}}")
+    @Value("#{${mosip.certify.authn.filter-urls}}")
     private List<String> urlPatterns;
 
     @Autowired
@@ -89,7 +89,7 @@ public class AccessTokenValidationFilter extends OncePerRequestFilter {
                     Jwt jwt = getNimbusJwtDecoder().decode(token);
                     parsedAccessToken.setClaims(new HashMap<>());
                     parsedAccessToken.getClaims().putAll(jwt.getClaims());
-                    parsedAccessToken.setAccessTokenHash(IdentityProviderUtil.generateOIDCAtHash(token));
+                    parsedAccessToken.setAccessTokenHash(CommonUtil.generateOIDCAtHash(token));
                     parsedAccessToken.setActive(true);
                     filterChain.doFilter(request, response);
                     return;
