+++ certify-service/src/main/java/io/mosip/certify/proof/ProofValidator.java
@@ -3,9 +3,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package io.mosip.certify.vci.pop;
+package io.mosip.certify.proof;
 
-import io.mosip.certify.core.dto.vci.CredentialProof;
+import io.mosip.certify.core.dto.CredentialProof;
 
 public interface ProofValidator {
