+++ certify-service/src/main/resources/application-local.properties
@@ -7,10 +7,10 @@ mosip.certify.security.auth.get-urls={}
 
 mosip.certify.security.ignore-csrf-urls=**/actuator/**,/favicon.ico,**/error,\
   **/swagger-ui/**,**/v3/api-docs/**,\
-  **/vci/**
+  **/issuance/**
 
 mosip.certify.security.ignore-auth-urls=**/actuator/**,**/error,**/swagger-ui/**,\
-  **/v3/api-docs/**, **/vci/**
+  **/v3/api-docs/**, **/issuance/**
 
 
 ## ------------------------------------------ Discovery openid-configuration -------------------------------------------
@@ -18,20 +18,20 @@ mosipbox.public.url=http://localhost:8090
 mosip.certify.discovery.issuer-id=${mosipbox.public.url}${server.servlet.path}
 
 ##--------------change this later---------------------------------
-mosip.certify.vci.supported.jwt-proof-alg={'RS256','PS256'}
+mosip.certify.supported.jwt-proof-alg={'RS256','PS256'}
 
 ##  ---------------------------------------------- VCI ------------------------------------------------------------------
 ##----- These are properties for any oauth resource server providing jwk------------###
-mosip.certify.vci.identifier=http://localhost:8088
-mosip.certify.vci.authn.filter-urls={ '${server.servlet.path}/vci/credential' }
-mosip.certify.vci.authn.issuer-uri=http://localhost:8088/v1/esignet
-mosip.certify.vci.authn.jwk-set-uri=http://localhost:8088/v1/esignet/oauth/.well-known/jwks.json
-mosip.certify.vci.authn.allowed-audiences={ '${mosipbox.public.url}${server.servlet.path}/vci/credential', 'http://localhost:8088/v1/esignet/vci/credential' }
+mosip.certify.identifier=http://localhost:8088
+mosip.certify.authn.filter-urls={ '${server.servlet.path}/issuance/credential' }
+mosip.certify.authn.issuer-uri=http://localhost:8088/v1/esignet
+mosip.certify.authn.jwk-set-uri=http://localhost:8088/v1/esignet/oauth/.well-known/jwks.json
+mosip.certify.authn.allowed-audiences={ '${mosipbox.public.url}${server.servlet.path}/issuance/credential', 'http://localhost:8088/v1/esignet/vci/credential' }
 
-mosip.certify.vci.key-values={\
+mosip.certify.key-values={\
  'v11' : {\
               'credential_issuer': '${mosipbox.public.url}',   \
-              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/vci/credential', \
+              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/issuance/credential', \
               'display': {{'name': 'Insurance', 'locale': 'en'}},\
               'credentials_supported': {{\
                       'format': 'ldp_vc',\
@@ -92,7 +92,7 @@ mosip.certify.vci.key-values={\
           },\
    'latest' : {\
               'credential_issuer': '${mosipbox.public.url}',   \
-              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/vci/credential', \
+              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/issuance/credential', \
               'display': {{'name': 'Insurance', 'locale': 'en'}},\
               'credentials_supported' : { \
                  "InsuranceCredential" : {\
@@ -156,13 +156,6 @@ mosip.certify.vci.key-values={\
 mosip.certify.integration.scan-base-package=io.mosip.certify.sunbirdrc.integration
 mosip.certify.integration.vci-plugin=SunbirdRCVCIssuancePlugin
 mosip.certify.integration.audit-plugin=LoggerAuditService
-
-##--------------------sunbird registry authentication related demo configuration-------------------------##
-
-mosip.certify.authenticator.sunbird-rc.auth-factor.kba.individual-id-field=policyNumber
-mosip.certify.authenticator.sunbird-rc.auth-factor.kba.field-details={{"id":"policyNumber", "type":"text", "format":""},{"id":"fullName", "type":"text", "format":""},{"id":"dob", "type":"date", "format":"dd/mm/yyyy"}}
-mosip.certify.authenticator.sunbird-rc.auth-factor.kba.registry-search-url=http://localhost:8000/registry/api/v1/Insurance/search
-mosip.certify.authenticator.sunbird-rc.kba.entity-id-field=osid
 
 ##-----------------------------VCI related demo configuration---------------------------------------------##
