+++ certify-service/src/test/resources/application-test.properties
@@ -18,18 +18,18 @@ spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.security.ser
 ##  ---------------------------------------------- VCI ------------------------------------------------------------------
 
 
-mosip.certify.vci.identifier=${mosipbox.public.url}${server.servlet.path}
-mosip.certify.vci.authn.filter-urls={ '${server.servlet.path}/vci/credential' }
-mosip.certify.vci.authn.issuer-uri=${mosipbox.public.url}${server.servlet.path}
-mosip.certify.vci.authn.jwk-set-uri=${mosipbox.public.url}${server.servlet.path}/oauth/.well-known/jwks.json
-mosip.certify.vci.authn.allowed-audiences={ '${mosipbox.public.url}${server.servlet.path}/vci/credential' }
+mosip.certify.identifier=${mosipbox.public.url}${server.servlet.path}
+mosip.certify.authn.filter-urls={ '${server.servlet.path}/issuance/credential' }
+mosip.certify.authn.issuer-uri=${mosipbox.public.url}${server.servlet.path}
+mosip.certify.authn.jwk-set-uri=${mosipbox.public.url}${server.servlet.path}/oauth/.well-known/jwks.json
+mosip.certify.authn.allowed-audiences={ '${mosipbox.public.url}${server.servlet.path}/issuance/credential' }
 
-mosip.certify.vci.supported.jwt-proof-alg={'RS256'}
+mosip.certify.supported.jwt-proof-alg={'RS256'}
 
-mosip.certify.vci.key-values={\
+mosip.certify.key-values={\
   "v11" : { \
               'credential_issuer': '${mosipbox.public.url}', 	\
-              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/vci/credential', \
+              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/issuance/credential', \
               'credentials_supported': {{\
               'format': 'ldp_vc',\
               'id': 'SampleVerifiableCredential_ldp', \
@@ -55,7 +55,7 @@ mosip.certify.vci.key-values={\
             },\
   "latest" : { \
               'credential_issuer': '${mosipbox.public.url}', 	\
-              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/vci/credential', \
+              'credential_endpoint': '${mosipbox.public.url}${server.servlet.path}/issuance/credential', \
               'display': {{'name': 'e-Signet', 'locale': 'en'}},\
               'credentials_supported' : { \
                  "SampleVerifiableCredential_ldp" : {\
