--- vci-service-impl/pom.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.mosip.certify</groupId>
-        <artifactId>certify-parent</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-
-    <groupId>io.mosip.certify</groupId>
-    <artifactId>vci-service-impl</artifactId>
-    <version>${project.parent.version}</version>
-
-    <properties>
-        <certify.core.version>${project.version}</certify.core.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.mosip.certify</groupId>
-            <artifactId>certify-core</artifactId>
-            <version>${certify.core.version}</version>
-        </dependency>
-    </dependencies>
-</project>
